+++ changelog.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.14.0] - Upcoming soon
+## [0.14.0] - 2020-07-26
 ### Changed
 * Removed workaround for deployment of Azure Integration Runtimes.
 ### Added